+++ forest/cli/src/cli/snapshot_cmd.rs
@@ -11,9 +11,9 @@ use forest_chain::ChainStore;
 use forest_cli_shared::cli::{
     default_snapshot_dir, is_car_or_tmp, snapshot_fetch, SnapshotServer, SnapshotStore,
 };
-use forest_db::db_engine::open_db;
+use forest_db::db_engine::{db_root, open_proxy_db};
 use forest_genesis::{forest_load_car, read_genesis_header};
-use forest_ipld::{recurse_links_hash, CidHashSet};
+use forest_ipld::{recurse_links_hash, CidHashSet, DEFAULT_RECENT_STATE_ROOTS};
 use forest_rpc_api::chain_api::ChainExportParams;
 use forest_rpc_client::chain_ops::*;
 use forest_utils::{io::parser::parse_duration, net::FetchProgress, retry};
@@ -35,12 +35,6 @@ pub(crate) const OUTPUT_PATH_DEFAULT_FORMAT: &str =
 pub enum SnapshotCommands {
     /// Export a snapshot of the chain to `<output_path>`
     Export {
-        /// Tipset to start the export from, default is the chain head
-        #[arg(short, long)]
-        tipset: Option<i64>,
-        /// Specify the number of recent state roots to include in the export.
-        #[arg(short, long, default_value = "2000")]
-        recent_stateroots: i64,
         /// Snapshot output path. Default to
         /// `forest_snapshot_{chain}_{year}-{month}-{day}_height_{height}.car`
         /// Date is in ISO 8601 date format.
@@ -150,8 +144,6 @@ impl SnapshotCommands {
     pub async fn run(&self, config: Config) -> anyhow::Result<()> {
         match self {
             Self::Export {
-                tipset,
-                recent_stateroots,
                 output_path,
                 skip_checksum,
                 dry_run,
@@ -161,7 +153,7 @@ impl SnapshotCommands {
                     Err(_) => cli_error_and_die("Could not get network head", 1),
                 };
 
-                let epoch = tipset.unwrap_or(chain_head.epoch());
+                let epoch = chain_head.epoch();
 
                 let now = OffsetDateTime::now_utc();
 
@@ -196,7 +188,7 @@ impl SnapshotCommands {
 
                 let params = ChainExportParams {
                     epoch,
-                    recent_roots: *recent_stateroots,
+                    recent_roots: DEFAULT_RECENT_STATE_ROOTS,
                     output_path,
                     tipset_keys: TipsetKeysJson(chain_head.key().clone()),
                     skip_checksum: *skip_checksum,
@@ -393,8 +385,13 @@ async fn validate(
 
     if confirm {
         let tmp_chain_data_path = TempDir::new()?;
-        let db_path = tmp_chain_data_path.path().join(&config.chain.name);
-        let db = open_db(&db_path, config.db_config())?;
+        let db_path = db_root(
+            tmp_chain_data_path
+                .path()
+                .join(&config.chain.name)
+                .as_path(),
+        );
+        let db = open_proxy_db(db_path, config.db_config().clone())?;
 
         let genesis = read_genesis_header(
             config.client.genesis_file.as_ref(),
@@ -413,7 +410,7 @@ async fn validate(
         let cids = {
             let file = tokio::fs::File::open(&snapshot).await?;
             let reader = FetchProgress::fetch_from_file(file).await?;
-            forest_load_car(chain_store.blockstore(), reader.compat()).await?
+            forest_load_car(chain_store.blockstore().clone(), reader.compat()).await?
         };
 
         let ts = chain_store.tipset_from_keys(&TipsetKeys::new(cids))?;