+++ ipld/src/util.rs
@@ -1,10 +1,11 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use std::future::Future;
+use std::{collections::VecDeque, future::Future};
 
 use cid::Cid;
-use fvm_ipld_encoding::from_slice;
+use forest_blocks::{BlockHeader, Tipset};
+use fvm_ipld_encoding::{from_slice, Cbor};
 
 use crate::{CidHashSet, Ipld};
 
@@ -77,4 +78,71 @@ where
     traverse_ipld_links_hash(walked, load_block, &ipld).await?;
 
     Ok(())
+}
+
+pub const DEFAULT_RECENT_STATE_ROOTS: i64 = 2000;
+
+/// Walks over tipset and state data and loads all blocks not yet seen.
+/// This is tracked based on the callback function loading blocks.
+pub async fn walk_snapshot<F, T>(
+    tipset: &Tipset,
+    recent_roots: i64,
+    mut load_block: F,
+) -> anyhow::Result<()>
+where
+    F: FnMut(Cid) -> T + Send,
+    T: Future<Output = anyhow::Result<Vec<u8>>> + Send,
+{
+    let mut seen = CidHashSet::default();
+    let mut blocks_to_walk: VecDeque<Cid> = tipset.cids().to_vec().into();
+    let mut current_min_height = tipset.epoch();
+    let incl_roots_epoch = tipset.epoch() - recent_roots;
+
+    while let Some(next) = blocks_to_walk.pop_front() {
+        if !seen.insert(&next) {
+            continue;
+        }
+
+        let data = load_block(next).await?;
+
+        let h = BlockHeader::unmarshal_cbor(&data)?;
+
+        if current_min_height > h.epoch() {
+            current_min_height = h.epoch();
+        }
+
+        if h.epoch() > incl_roots_epoch {
+            recurse_links_hash(&mut seen, *h.messages(), &mut load_block).await?;
+        }
+
+        if h.epoch() > 0 {
+            for p in h.parents().cids() {
+                blocks_to_walk.push_back(*p);
+            }
+        } else {
+            for p in h.parents().cids() {
+                load_block(*p).await?;
+            }
+        }
+
+        if h.epoch() == 0 || h.epoch() > incl_roots_epoch {
+            recurse_links_hash(&mut seen, *h.state_root(), &mut load_block).await?;
+        }
+    }
+
+    Ok(())
+}
+
+pub fn should_save_block_to_snapshot(cid: &Cid) -> bool {
+    // Don't include identity CIDs.
+    // We only include raw and dagcbor, for now.
+    // Raw for "code" CIDs.
+    if cid.hash().code() == u64::from(cid::multihash::Code::Identity) {
+        false
+    } else {
+        matches!(
+            cid.codec(),
+            fvm_shared::IPLD_RAW | fvm_ipld_encoding::DAG_CBOR
+        )
+    }
 }