+++ node/db/src/parity_db.rs
@@ -50,8 +50,8 @@ impl ParityDb {
         })
     }
 
-    pub fn open(path: PathBuf, config: &ParityDbConfig) -> anyhow::Result<Self> {
-        let opts = Self::to_options(path, config)?;
+    pub fn open(path: impl Into<PathBuf>, config: &ParityDbConfig) -> anyhow::Result<Self> {
+        let opts = Self::to_options(path.into(), config)?;
         Ok(Self {
             db: Arc::new(Db::open_or_create(&opts)?),
             statistics_enabled: opts.stats,
@@ -104,14 +104,6 @@ impl Store for ParityDb {
         //     }))
         // }
         // ```
-    }
-
-    fn delete<K>(&self, key: K) -> Result<(), Error>
-    where
-        K: AsRef<[u8]>,
-    {
-        let tx = [(0, key.as_ref(), None)];
-        self.db.commit(tx).map_err(Error::from)
     }
 
     fn exists<K>(&self, key: K) -> Result<bool, Error>