+++ scripts/gen_coverage_report.sh
@@ -35,6 +35,8 @@ cov forest --chain calibnet --encrypt-keystore false --import-snapshot "$SNAPSHO
 cov forest-cli sync wait
 cov forest-cli sync status
 cov forest-cli chain validate-tipset-checkpoints
+cov forest-cli --chain calibnet db gc
+cov forest-cli --chain calibnet db stats
 cov forest-cli snapshot export
 cov forest-cli attach --exec 'showPeers()'
 cov forest-cli net listen