+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Notable updates:
 - [forest daemon] Support for NV18.
   [#2558](https://github.com/ChainSafe/forest/pull/2558)
   [#2579](https://github.com/ChainSafe/forest/pull/2579)
+- [forest daemon] Automatic database garbage collection.
+  [#2638](https://github.com/ChainSafe/forest/pull/2638)
 
 ### Changed